issue solved

+++ src/main/java/unicam/model/iscrizione/Iscrizione.java
@@ -5,33 +5,33 @@ import unicam.model.utenti.user.User;
 import java.util.List;
 
 public class Iscrizione {
-    private int id;
-    private int htId;
-    private int teamId;
+    private Long id;
+    private Long htId;
+    private Long teamId;
 
     public Iscrizione() {}
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getHtId() {
+    public Long getHtId() {
         return htId;
     }
 
-    public void setHtId(int htId) {
+    public void setHtId(Long htId) {
         this.htId = htId;
     }
 
-    public int getTeamId() {
+    public Long getTeamId() {
         return teamId;
     }
 
-    public void setTeamId(int teamId) {
+    public void setTeamId(Long teamId) {
         this.teamId = teamId;
     }
 }
